test: separate out all cos tests
