+++ .github/workflows/integration_test.yaml
@@ -15,7 +15,7 @@ jobs:
         --openstack-rc ${GITHUB_WORKSPACE}/openrc
         --kube-config ${GITHUB_WORKSPACE}/kube-config
         --screenshot-dir /tmp
-      modules: '["test_addon", "test_core", "test_cos", "test_ingress"]'
+      modules: '["test_addon", "test_core", "test_ingress", "test_cos_grafana", "test_cos_loki", "test_cos_prometheus"]'
       pre-run-script: |
         -c "sudo microk8s enable hostpath-storage
           sudo microk8s kubectl -n kube-system rollout status -w deployment/hostpath-provisioner
